move middlewares to outside te controller folder
