+++ server/middleware/index.js
@@ -1,4 +1,5 @@
 const isAuth = require('./isAuth');
+
 module.exports = {
-    isAuth
-}
+  isAuth,
+};
