+++ server/utils/index.js
@@ -1,5 +1,6 @@
 const { signToken, verifyToken } = require('./tokenFunction');
+
 module.exports = {
-    signToken,
-    verifyToken
-}
+  signToken,
+  verifyToken,
+};
